backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.3
Flask-CORS==3.0.10
Flask-WTF==1.0.0
SQLAlchemy==1.4.41
pymysql==1.0.2
python-dotenv==0.20.0
passlib==1.7.4
requests==2.28.1
pandas==1.4.4
openpyxl==3.0.10
redis==4.3.4
